restructure buildtools
